requirements.txt
Flask-SQLAlchemy>=1.0
pytest>=2.5.2
